bug8

+++ vlib/picoev/multiplex_mac.v
@@ -64,7 +64,7 @@ pub fn (mut pv Picoev) ev_set(fd int, operation int, events int) {
 
 [inline]
 fn backend_build(next_fd int, events int) int {
-	return ((next_fd << 8) | (events & 0xff))
+	return int((u32(next_fd) << 8) | (events & 0xff)))
 }
 
 [inline]
